third_party/blink/web_tests/http/tests/security/isolatedWorld/websocket/isolated-world-websocket-csp-expected.txt

CONSOLE MESSAGE: Testing main world. Request should be blocked by main world CSP.
CONSOLE ERROR: Connecting to 'ws://127.0.0.1:8880/echo' violates the following Content Security Policy directive: "connect-src 'none'". The action has been blocked.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
CONSOLE MESSAGE: Testing isolated world with no csp. Request should be blocked by main world CSP.
CONSOLE ERROR: Connecting to 'ws://127.0.0.1:8880/echo' violates the following Content Security Policy directive: "connect-src 'none'". The action has been blocked.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
CONSOLE MESSAGE: Testing isolated world with permissive csp.
CONSOLE MESSAGE: PASS: Request succeeded as expected.
CONSOLE MESSAGE: Testing isolated world with strict csp.
CONSOLE ERROR: Connecting to 'ws://127.0.0.1:8880/echo' violates the following Content Security Policy directive: "connect-src 'none'". The action has been blocked.
CONSOLE MESSAGE: PASS: Request blocked by CSP as expected.
This tests the interaction of websockets used in the isolated world with the isolated world CSP.
